refactored document indexing
- use tika to determine content-type
- refactored document creation into a document builder class
- use this class in the index build and update_handler
