+++ lib/mu_search/automatic_update_handler.rb
@@ -40,10 +40,12 @@ module MuSearch
         indexes = Indexes.instance.get_indexes(index_type)
         indexes.each do |key, index|
           allowed_groups = index[:allowed_groups]
+          sparql_client = build_sparql_client(allowed_groups)
+          document_builder = DocumentBuilder.new(tika_client: @tika_client, sparql_client: sparql_client)
           @logger.debug "Got allowed groups for updated_document_all_types #{allowed_groups}"
           rdf_type = @type_definitions.dig(index_type, "rdf_type")
           @logger.debug "Got RDF type for updated_document_all_types #{rdf_type}"
-          if document_exists_for(document_id, rdf_type, allowed_groups)
+          if document_exists_for(client, document_id, rdf_type)
             @logger.debug "Our current index knows that #{document_id} is of type #{rdf_type} based on allowed groups #{allowed_groups}"
             properties = @type_definitions.dig(index_type, "properties")
             document = fetch_document_to_index @tika_client, uri: document_id, properties: properties,
@@ -72,8 +74,9 @@ module MuSearch
         indexes = Indexes.instance.get_indexes(index_type)
         indexes.each do |key, index|
           allowed_groups = index[:allowed_groups]
+          sparql_client = build_sparql_client(allowed_groups)
           type = @type_definitions.dig(index_type, "rdf_type")
-          if document_exists_for(document_id, type, allowed_groups) == "true"
+          if document_exists_for(client, document_id, type) == "true"
             @logger.debug "Not deleting document #{document_id} from #{index[:index]}, it still exists"
           else
             @logger.debug "Deleting document #{document_id} from #{index[:index]}"
@@ -85,6 +88,13 @@ module MuSearch
           end
         end
       end
+    end
+
+    private
+    def build_sparql_client
+      allowed_groups_object = allowed_groups.select { |group| group }
+      sparql_options = { headers: { 'mu-auth-allowed-groups': allowed_groups_object.to_json } }
+      ::SPARQL::Client.new(ENV['MU_SPARQL_ENDPOINT'], sparql_options)
     end
   end
 end
