+++ web.rb
@@ -12,6 +12,7 @@ require_relative 'lib/mu_search/delta_handler.rb'
 require_relative 'lib/mu_search/automatic_update_handler.rb'
 require_relative 'lib/mu_search/invalidating_update_handler.rb'
 require_relative 'lib/mu_search/config_parser.rb'
+require_relative 'lib/mu_search/document_builder.rb'
 require_relative 'lib/mu_search/index_builder.rb'
 require_relative 'framework/elastic.rb'
 require_relative 'framework/tika.rb'
